move structs out
